Extract form to other component
